Separate memory ports into their own file
